Move components under lib package and add compatibility shim
