+++ docs/implementations/data-quality-governance/collibra.md
@@ -70,8 +70,8 @@ A Collibra-backed `DQClient` typically combines the existing Collibra adapter
 used for contract resolution with calls that manage compatibility assets:
 
 ```python
-from dc43.components.contract_store import HttpCollibraContractAdapter
-from dc43.components.data_quality.governance import DQStatus
+from dc43.lib.components.contract_store import HttpCollibraContractAdapter
+from dc43.lib.components.data_quality.governance import DQStatus
 
 class CollibraDQClient:
     def __init__(self, adapter: HttpCollibraContractAdapter, asset_type: str):
