+++ src/dc43/components/__init__.py
@@ -1,18 +1,24 @@
-"""Composable dc43 components grouped by lifecycle area.
+"""Compatibility shim that proxies to :mod:`dc43.lib.components`."""
+from __future__ import annotations
 
-This package surfaces high-level building blocks (contract stores, drafting,
-data-quality governance, runtime integrations) alongside their reference
-implementations. Subpackages are structured consistently so it is easier to
-swap implementations without scanning unrelated helpers.
-"""
+import importlib
+from typing import Any, Iterable
 
-from . import contract_store, contract_drafter, data_quality, integration
-from .contract_store import ContractStore
+_lib_components = importlib.import_module("dc43.lib.components")
 
-__all__ = [
-    "ContractStore",
-    "contract_store",
-    "contract_drafter",
-    "data_quality",
-    "integration",
-]
+# Expose the same package search path so submodules continue to resolve when
+# imported via the legacy location (e.g. ``dc43.components.data_quality``).
+__path__ = _lib_components.__path__  # type: ignore[attr-defined]
+
+
+def __getattr__(name: str) -> Any:
+    return getattr(_lib_components, name)
+
+
+def __dir__() -> Iterable[str]:
+    return getattr(_lib_components, "__all__", [])
+
+
+from dc43.lib.components import *  # noqa: F401,F403 - re-export public API
+
+__all__ = getattr(_lib_components, "__all__", [])
