+++ src/dc43/lib/components/__init__.py
@@ -1,0 +1,18 @@
+"""Composable dc43 components grouped by lifecycle area.
+
+This package surfaces high-level building blocks (contract stores, drafting,
+data-quality governance, runtime integrations) alongside their reference
+implementations. Subpackages are structured consistently so it is easier to
+swap implementations without scanning unrelated helpers.
+"""
+
+from . import contract_store, contract_drafter, data_quality, integration
+from .contract_store import ContractStore
+
+__all__ = [
+    "ContractStore",
+    "contract_store",
+    "contract_drafter",
+    "data_quality",
+    "integration",
+]
