+++ src/dc43/lib/components/contract_store/impl/delta.py
