+++ src/dc43/lib/components/contract_validation/core.py
