+++ src/dc43/services/data_quality/client/interface.py
@@ -6,7 +6,7 @@ from typing import Protocol
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.components.contract_validation import ValidationResult
+from dc43.lib.components.contract_validation import ValidationResult
 from dc43.lib.data_quality import ObservationPayload
