+++ src/dc43/services/data_quality/client/local.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.components.contract_validation import ValidationResult
+from dc43.lib.components.contract_validation import ValidationResult
 from dc43.lib.data_quality import ObservationPayload
 
 from ..backend import DataQualityServiceBackend, LocalDataQualityServiceBackend
