+++ src/dc43/services/governance/backend/interface.py
@@ -6,8 +6,8 @@ from typing import Callable, Mapping, Optional, Protocol, Sequence
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.components.contract_validation import ValidationResult
-from dc43.components.data_quality.governance import DQStatus
+from dc43.lib.components.contract_validation import ValidationResult
+from dc43.lib.components.data_quality.governance import DQStatus
 from dc43.lib.data_quality import ObservationPayload
 
 from ..models import (
