+++ src/dc43/services/governance/client/local.py
@@ -6,9 +6,9 @@ from typing import Callable, Mapping, Optional, Sequence
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.components.contract_store.interface import ContractStore
-from dc43.components.contract_validation import ValidationResult
-from dc43.components.data_quality.governance import DQStatus
+from dc43.lib.components.contract_store.interface import ContractStore
+from dc43.lib.components.contract_validation import ValidationResult
+from dc43.lib.components.data_quality.governance import DQStatus
 from dc43.lib.data_quality import ObservationPayload
 from dc43.services.contracts.backend import (
     ContractServiceBackend,
