+++ tests/test_collibra_integration.py
@@ -5,11 +5,11 @@ from datetime import datetime
 
 import pytest
 
-from dc43.components.contract_store import (
+from dc43.lib.components.contract_store import (
     HttpCollibraContractAdapter,
     StubCollibraContractAdapter,
 )
-from dc43.components.contract_store.impl.collibra import CollibraContractStore
+from dc43.lib.components.contract_store.impl.collibra import CollibraContractStore
 from dc43.odcs import build_odcs
 from open_data_contract_standard.model import SchemaProperty, SchemaObject, Server  # type: ignore
